make post ripples appear more consistently across app
posts that appear on the home tab already have ripples.
this also adds ripples on posts that appear in the search and
notification tabs, in replies, and in embedded quotes.

ripples on posts that appear in replies look kinda ugly  (especially in
threaded mode) because the pressable area doesn't extend horizontally to
the edges of the screen, not sure if something should be done about this